リダイレクトがきた時のファイルのopenと作成部分を実装。細かいバグを修正。
