+++ Makefile
@@ -4,6 +4,8 @@ LEXER_DIR = ./srcs/lexer/
 
 PARSER_DIR = ./srcs/parse/
 
+EXPANSION_DIR = ./srcs/expansion/
+
 BUILTIN_DIR = ./srcs/builtin_cmd/
 
 UTILS_DIR = ./srcs/utils/
@@ -11,6 +13,8 @@ UTILS_DIR = ./srcs/utils/
 SRCS_LEXER = lexer.c is_function.c lst_function.c free_function.c init_datas.c
 
 SRCS_PARSER = parse.c init_lst_function.c parse_lst_function.c create_lst.c free_node.c
+
+SRCS_EXPANSION = expansion.c hear_doc.c delete_quote.c sepalate_str.c
 
 SRCS_BUILTIN = cd.c echo.c env.c exit.c export.c lst_func.c pwd.c unset.c
 
@@ -20,11 +24,13 @@ SRCS_L = ${addprefix ${LEXER_DIR}, ${SRCS_LEXER}}
 
 SRCS_P = ${addprefix ${PARSER_DIR}, ${SRCS_PARSER}}
 
+SRCS_E = ${addprefix ${EXPANSION_DIR}, ${SRCS_EXPANSION}}
+
 SRCS_B = ${addprefix ${BUILTIN_DIR}, ${SRCS_BUILTIN}}
 
 SRCS_U = ${addprefix ${UTILS_DIR}, ${SRCS_UTILS}}
 
-SRCS_NAME = ./srcs/main.c ${SRCS_L} ${SRCS_P} ${SRCS_B} ${SRCS_U} ./srcs/expansion/expansion.c ./srcs/expansion/hear_doc.c ./srcs/exection/exection.c ./get_next_line/get_next_line.c ./get_next_line/get_next_line_utils.c
+SRCS_NAME = ./srcs/main.c ${SRCS_L} ${SRCS_P} ${SRCS_E} ${SRCS_B} ${SRCS_U} ./srcs/exection/exection.c ./get_next_line/get_next_line.c ./get_next_line/get_next_line_utils.c
 
 HEADER = ./includes
